// src/enums/bomb-action-types.enum.ts
export enum BOMB_ACTION_TYPES {
    RESET_STORE = '[bombs] RESET STORE',
    EMIT_NEW_BOMB = '[bombs] EMIT NEW BOMB',
    REMOVE_BOMB = '[bombs] REMOVE BOMB',
    SET_DRAGGED_BOMB_ID = '[bombs] SET DRAGGED BOMB ID',
    SET_SCORE = '[bombs] SET SCORE',
    SWAP_BIN_COLORS = '[bombs] SWAP BIN COLORS',
    DECREMENT_BOMBS_TIMER = '[bombs] DECREMENT BOMBS TIMER',
    SET_INTERSECTED_BIN_ID = '[bombs] SET INTERSECTED BIN ID',
    SET_GAME_STATUS = '[bombs] SET GAME STATUS'
}